pyproject.toml
[tool.poetry]
name = "slack-machine-plugin-rbac"
version = "0.1.0"
license = "MIT"
description = "A Slack-Machine plugin for role based access control"
authors = ["David Jack Wange Olrik <david@olrik.dk>"]
readme = "README.md"
homepage = "https://github.com/davidolrik/slack-machine-plugin-rbac/"
repository = "https://github.com/davidolrik/slack-machine-plugin-rbac/"
keywords = ['slack', 'bot', 'plugin']
classifiers = [
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
]
packages = [
    { include = "slack_machine_plugin_rbac/*.py", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7.0"
slack-machine = "^0.18.2"
arrow = "^0.15.4"

# [tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
